requirements.txt
streamlit
Pillow
PyMuPDF
langchain
openai
tiktoken
faiss-cpu
google-generativeai
python-dotenv
pdf2image
langchain
streamlit
ipykernel
python-dotenv
langchain-openai
langchain_groq
langchain_core
fastapi
uvicorn
langserve
langchain_community
pypdf
bs4
langchain_chroma
langchain_huggingface
selenium
pymongo